+++ packages/eslint-plugin/tests/rules/no-inputs-metadata-property/cases.ts
@@ -45,6 +45,41 @@ export const valid = [
     })
     class Test {}
   `,
+  /**
+   * Using inputs when using the directive composition API is not a bad practice
+   * https://angular.io/guide/directive-composition-api
+   * https://www.youtube.com/watch?v=EJJwyyjsRGs
+   */
+  `
+   @Component({
+     selector: 'qx-menuitem',
+     hostDirectives: [{
+       directive: CdkMenuItem,
+       inputs: ['cdkMenuItemDisabled: disabled'],
+     }]
+   })
+   class Test {}
+ `,
+  `
+   @Component({
+     selector: 'qx-menuitem',
+     'hostDirectives': [{
+       directive: CdkMenuItem,
+       inputs: ['cdkMenuItemDisabled: disabled'],
+     }]
+   })
+   class Test {}
+ `,
+  `
+   @Component({
+     selector: 'qx-menuitem',
+     ['hostDirectives']: [{
+       directive: CdkMenuItem,
+       inputs: ['cdkMenuItemDisabled: disabled'],
+     }]
+   })
+   class Test {}
+ `,
 ];
 
 export const invalid = [